Shuffle stuff around + Knex config'd
