--- app.js
@@ -1,123 +1,0 @@
-/**
- * Module dependencies.
- */
-var express = require('express');
-var cookieParser = require('cookie-parser');
-var compress = require('compression');
-var favicon = require('serve-favicon');
-var session = require('express-session');
-var bodyParser = require('body-parser');
-var logger = require('morgan');
-var errorHandler = require('errorhandler');
-var lusca = require('lusca');
-var methodOverride = require('method-override');
-
-var _ = require('lodash');
-var flash = require('express-flash');
-var path = require('path');
-var passport = require('passport');
-var expressValidator = require('express-validator');
-
-/**
- * Controllers (route handlers).
- */
-var homeController = require('./controllers/home');
-var userController = require('./controllers/user');
-var apiController = require('./controllers/api');
-var apiExamplesController = require('./controllers/api-examples');
-var contactController = require('./controllers/contact');
-var socketController = require('./controllers/socket');
-/**
- * API keys and Passport configuration.
- */
-var secrets = require('./config/secrets');
-var passportConf = require('./config/passport');
-
-/**
- * Create Express server.
- */
-var app = express();
-var http = require('http').Server(app);
-var io = require('socket.io')(http);
-
-/**
- * Express configuration.
- */
-app.set('port', process.env.PORT || 3000);
-app.set('views', path.join(__dirname, 'views'));
-app.set('view engine', 'jade');
-app.use(compress());
-app.use(logger('dev'));
-app.use(favicon(path.join(__dirname, 'public', 'favicon.png')));
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(expressValidator());
-app.use(methodOverride());
-app.use(cookieParser());
-app.use(session({
-  resave: true,
-  saveUninitialized: true,
-  secret: secrets.sessionSecret
-}));
-app.use(passport.initialize());
-app.use(passport.session());
-app.use(flash());
-app.use(lusca({
-  csrf: false,
-  xframe: 'SAMEORIGIN',
-  xssProtection: true
-}));
-app.use(function(req, res, next) {
-  res.locals.user = req.user;
-  next();
-});
-app.use(function(req, res, next) {
-  if (/api/i.test(req.path)) req.session.returnTo = req.path;
-  next();
-});
-app.use(express.static(path.join(__dirname, 'public'), { maxAge: 31557600000 }));
-
-/**
- *  Socket connection endpoints
- */
-socketController.setIo(io);
-
-/**
- * Primary app routes.
- */
-app.get('/', homeController.index);
-app.get('/login', userController.getLogin);
-app.post('/login', userController.postLogin);
-app.get('/logout', userController.logout, socketController.logout);
-app.get('/forgot', userController.getForgot);
-app.post('/forgot', userController.postForgot);
-app.get('/reset/:token', userController.getReset);
-app.post('/reset/:token', userController.postReset);
-app.get('/signup', userController.getSignup);
-app.post('/signup', userController.postSignup);
-app.get('/contact', contactController.getContact);
-app.post('/contact', contactController.postContact);
-app.get('/account', passportConf.isAuthenticated, userController.getAccount);
-app.post('/account/profile', passportConf.isAuthenticated, userController.postUpdateProfile);
-app.post('/account/password', passportConf.isAuthenticated, userController.postUpdatePassword);
-app.post('/account/delete', passportConf.isAuthenticated, userController.postDeleteAccount);
-app.get('/account/unlink/:provider', passportConf.isAuthenticated, userController.getOauthUnlink);
-
-/**
- * API Routes
- */
-app.post('/api/login', apiController.postLogin);
-
-/**
- * Error Handler.
- */
-app.use(errorHandler());
-
-/**
- * Start Express server.
- */
-http.listen(app.get('port'), function(){
-  console.log('Express server listening on port %d in %s mode', app.get('port'), app.get('env'));
-});
-
-module.exports = app;
