+++ package.json
@@ -10,13 +10,14 @@
     "npm": "3.10.8"
   },
   "scripts": {
-    "start": "node app.js",
+    "start": "nodemon ./src/app.js",
     "test": "mocha --reporter spec --timeout 5000"
   },
   "dependencies": {
     "async": "^1.5.0",
     "azure": "^0.10.6",
     "bcrypt-nodejs": "^0.0.3",
+    "bluebird": "^3.4.6",
     "body-parser": "^1.14.1",
     "cfenv": "1.0.3",
     "compression": "^1.6.0",
@@ -53,6 +54,7 @@
     "grunt": "~0.4.5",
     "mocha": "^2.3.3",
     "multiline": "^1.0.2",
+    "nodemon": "^1.11.0",
     "supertest": "^1.1.0"
   }
 }
