+++ src/constants/tables.json
@@ -1,0 +1,3 @@
+{
+  "USERS":"users"
+}
